move modal out into own component
