+++ protocol-designer/src/organisms/index.ts
@@ -1,3 +1,4 @@
+export * from './FeatureFlagsModal'
 export * from './FileUploadMessagesModal/'
 export * from './IncompatibleTipsModal'
 export * from './Kitchen'
